homework2/movie_theater_booking/requirements.txt
Django==4.2.19
djangorestframework==3.14.0
django-widget-tweaks==1.5.0 